Moves all the gates to fock_gradient
