+++ thewalrus/tests/test_fock_gradients2.py
@@ -19,73 +19,3 @@ import numpy as np
 np.set_printoptions(linewidth=100)
 
 
-def test_Dgate():
-    """Tests the value of the analytic gradient for the Dgate against finite differences"""
-    cutoff = 4
-    r = 1.0
-    theta = np.pi / 8
-    _, Dr, Dtheta = Dgate(r, theta, cutoff, grad=True)
-    dr = 0.001
-    dtheta = 0.001
-    Drp, _, _ = Dgate(r + dr, theta, cutoff, grad=False)
-    Drm, _, _ = Dgate(r - dr, theta, cutoff, grad=False)
-    Dthetap, _, _ = Dgate(r, theta + dtheta, cutoff, grad=False)
-    Dthetam, _, _ = Dgate(r, theta - dtheta, cutoff, grad=False)
-    Drapprox = (Drp - Drm) / (2 * dr)
-    Dthetaapprox = (Dthetap - Dthetam) / (2 * dtheta)
-    assert np.allclose(Dr, Drapprox, atol=1e-5, rtol=0)
-    assert np.allclose(Dtheta, Dthetaapprox, atol=1e-5, rtol=0)
-
-
-def test_Sgate():
-    """Tests the value of the analytic gradient for the Sgate against finite differences"""
-    cutoff = 4
-    r = 1.0
-    theta = np.pi / 8
-    _, Dr, Dtheta = Sgate(r, theta, cutoff, grad=True)
-    dr = 0.001
-    dtheta = 0.001
-    Drp, _, _ = Sgate(r + dr, theta, cutoff, grad=False)
-    Drm, _, _ = Sgate(r - dr, theta, cutoff, grad=False)
-    Dthetap, _, _ = Sgate(r, theta + dtheta, cutoff, grad=False)
-    Dthetam, _, _ = Sgate(r, theta - dtheta, cutoff, grad=False)
-    Drapprox = (Drp - Drm) / (2 * dr)
-    Dthetaapprox = (Dthetap - Dthetam) / (2 * dtheta)
-    assert np.allclose(Dr, Drapprox, atol=1e-5, rtol=0)
-    assert np.allclose(Dtheta, Dthetaapprox, atol=1e-5, rtol=0)
-
-
-def test_S2gate():
-    """Tests the value of the analytic gradient for the S2gate against finite differences"""
-    cutoff = 4
-    r = 1.0
-    theta = np.pi / 8
-    _, Dr, Dtheta = S2gate(r, theta, cutoff, grad=True)
-    dr = 0.001
-    dtheta = 0.001
-    Drp, _, _ = S2gate(r + dr, theta, cutoff, grad=False)
-    Drm, _, _ = S2gate(r - dr, theta, cutoff, grad=False)
-    Dthetap, _, _ = S2gate(r, theta + dtheta, cutoff, grad=False)
-    Dthetam, _, _ = S2gate(r, theta - dtheta, cutoff, grad=False)
-    Drapprox = (Drp - Drm) / (2 * dr)
-    Dthetaapprox = (Dthetap - Dthetam) / (2 * dtheta)
-    assert np.allclose(Dr, Drapprox, atol=1e-5, rtol=0)
-    assert np.allclose(Dtheta, Dthetaapprox, atol=1e-5, rtol=0)
-
-
-def test_BSgate():
-    """Tests the value of the analytic gradient for the BSgate against finite differences"""
-    cutoff = 4
-    r = 1.0
-    theta = np.pi / 8
-    _, Dr, Dtheta = BSgate(r, theta, cutoff, grad=True)
-    dr = 0.001
-    dtheta = 0.001
-    Drp, _, _ = BSgate(r + dr, theta, cutoff, grad=False)
-    Drm, _, _ = BSgate(r - dr, theta, cutoff, grad=False)
-    Dthetap, _, _ = BSgate(r, theta + dtheta, cutoff, grad=False)
-    Dthetam, _, _ = BSgate(r, theta - dtheta, cutoff, grad=False)
-    Drapprox = (Drp - Drm) / (2 * dr)
-    Dthetaapprox = (Dthetap - Dthetam) / (2 * dtheta)
-    assert np.allclose(Dr, Drapprox, atol=1e-4, rtol=0)
-    assert np.allclose(Dtheta, Dthetaapprox, atol=1e-4, rtol=0)
